linkedin_portfolio_generator/requirements.txt
requests
beautifulsoup4
selenium
openai
python-docx
pdfkit
python-dotenv 